chore: rename file extensions from web to native
