+++ src/hooks/useCSS.native.ts
@@ -1,0 +1,3 @@
+import type { ReactNativeStyleSheet } from '../types'
+
+export const useCSS = <T>(_stylesheet: ReactNativeStyleSheet<T>): void => {}
